--- app/src/main/java/woowacourse/shopping/data/datasource/ApiHandleOrderDataSource.kt
@@ -1,8 +1,0 @@
-package woowacourse.shopping.data.datasource
-
-import woowacourse.shopping.data.remote.api.ApiResponse
-import woowacourse.shopping.data.remote.dto.request.RequestOrderPostDto
-
-interface ApiHandleOrderDataSource {
-    suspend fun postOrder(request: RequestOrderPostDto): ApiResponse<Unit>
-}
