+++ app/src/main/java/woowacourse/shopping/data/datasource/CouponDataSource.kt
@@ -1,7 +1,8 @@
 package woowacourse.shopping.data.datasource
 
+import woowacourse.shopping.data.remote.api.ApiResponse
 import woowacourse.shopping.data.remote.dto.response.CouponDto
 
 interface CouponDataSource {
-    suspend fun getCoupons(): List<CouponDto>
+    suspend fun getCoupons(): ApiResponse<List<CouponDto>>
 }
