+++ app/src/main/java/woowacourse/shopping/data/datasource/ProductDataSource.kt
@@ -1,5 +1,6 @@
 package woowacourse.shopping.data.datasource
 
+import woowacourse.shopping.data.remote.api.ApiResponse
 import woowacourse.shopping.data.remote.dto.response.ResponseProductIdGetDto
 import woowacourse.shopping.data.remote.dto.response.ResponseProductsGetDto
 
@@ -7,7 +8,7 @@ interface ProductDataSource {
     suspend fun getProductsByOffset(
         page: Int,
         size: Int,
-    ): ResponseProductsGetDto
+    ): ApiResponse<ResponseProductsGetDto>
 
-    suspend fun getProductsById(id: Long): ResponseProductIdGetDto
+    suspend fun getProductsById(id: Long): ApiResponse<ResponseProductIdGetDto>
 }
