+++ app/src/main/java/woowacourse/shopping/data/datasource/impl/RemoteCartDataSource.kt
@@ -1,7 +1,9 @@
 package woowacourse.shopping.data.datasource.impl
 
 import woowacourse.shopping.data.datasource.CartDataSource
+import woowacourse.shopping.data.remote.api.ApiResponse
 import woowacourse.shopping.data.remote.api.ShoppingRetrofit
+import woowacourse.shopping.data.remote.api.handleApi
 import woowacourse.shopping.data.remote.dto.request.RequestCartItemPostDto
 import woowacourse.shopping.data.remote.dto.request.RequestCartItemsPatchDto
 import woowacourse.shopping.data.remote.dto.response.ResponseCartItemCountsGetDto
@@ -11,24 +13,23 @@ class RemoteCartDataSource : CartDataSource {
     override suspend fun getCartItems(
         page: Int,
         size: Int,
-    ): ResponseCartItemsGetDto =
-        ShoppingRetrofit.cartItemService.getCartItems(page, size).body() ?: error("body가 비어있습니다.")
+    ): ApiResponse<ResponseCartItemsGetDto> =
+        handleApi { ShoppingRetrofit.cartItemService.getCartItems(page = page, size = size) }
 
-    override suspend fun postCartItems(request: RequestCartItemPostDto) {
-        ShoppingRetrofit.cartItemService.postCartItem(request)
-    }
+    override suspend fun postCartItems(request: RequestCartItemPostDto): ApiResponse<Unit> =
+        handleApi { ShoppingRetrofit.cartItemService.postCartItem(request = request) }
 
-    override suspend fun deleteCartItems(id: Long) {
-        ShoppingRetrofit.cartItemService.deleteCartItem(id)
-    }
+    override suspend fun deleteCartItems(id: Long): ApiResponse<Unit> =
+        handleApi {
+            ShoppingRetrofit.cartItemService.deleteCartItem(id = id)
+        }
 
     override suspend fun patchCartItems(
         id: Long,
         request: RequestCartItemsPatchDto,
-    ) {
-        ShoppingRetrofit.cartItemService.patchCartItem(id, request)
-    }
+    ): ApiResponse<Unit> =
+        handleApi { ShoppingRetrofit.cartItemService.patchCartItem(id = id, request = request) }
 
-    override suspend fun getCartItemCounts(): ResponseCartItemCountsGetDto =
-        ShoppingRetrofit.cartItemService.getCartItemCounts().body() ?: error("body가 비어있습니다.")
+    override suspend fun getCartItemCounts(): ApiResponse<ResponseCartItemCountsGetDto> =
+        handleApi { ShoppingRetrofit.cartItemService.getCartItemCounts() }
 }
