+++ app/src/main/java/woowacourse/shopping/data/datasource/impl/RemoteCouponDataSource.kt
@@ -1,10 +1,14 @@
 package woowacourse.shopping.data.datasource.impl
 
 import woowacourse.shopping.data.datasource.CouponDataSource
+import woowacourse.shopping.data.remote.api.ApiResponse
 import woowacourse.shopping.data.remote.api.ShoppingRetrofit
+import woowacourse.shopping.data.remote.api.handleApi
 import woowacourse.shopping.data.remote.dto.response.CouponDto
 
 class RemoteCouponDataSource : CouponDataSource {
-    override suspend fun getCoupons(): List<CouponDto> =
-        ShoppingRetrofit.couponService.getCoupons().body() ?: emptyList()
+    override suspend fun getCoupons(): ApiResponse<List<CouponDto>> =
+        handleApi {
+            ShoppingRetrofit.couponService.getCoupons()
+        }
 }
