+++ app/src/main/java/woowacourse/shopping/data/datasource/impl/RemoteProductDataSource.kt
@@ -1,13 +1,13 @@
 package woowacourse.shopping.data.datasource.impl
 
-import woowacourse.shopping.data.datasource.ApiHandleProductDataSource
+import woowacourse.shopping.data.datasource.ProductDataSource
 import woowacourse.shopping.data.remote.api.ApiResponse
 import woowacourse.shopping.data.remote.api.ShoppingRetrofit
 import woowacourse.shopping.data.remote.api.handleApi
 import woowacourse.shopping.data.remote.dto.response.ResponseProductIdGetDto
 import woowacourse.shopping.data.remote.dto.response.ResponseProductsGetDto
 
-class ApiHandleProductDataSourceImpl : ApiHandleProductDataSource {
+class RemoteProductDataSource : ProductDataSource {
     override suspend fun getProductsByOffset(
         page: Int,
         size: Int,
